+++ apps/veil/src/entities/leaderboard/ui/page.tsx
@@ -1,8 +1,7 @@
 import { LeaderboardTable } from './table';
 import { PenumbraWaves } from '@/pages/explore/ui/waves';
 
-// eslint-disable-next-line @typescript-eslint/require-await -- next wills it
-export const LeaderboardPage = async () => {
+export const LeaderboardPage = () => {
   return (
     <section className='flex flex-col gap-6 p-4 max-w-[1062px] mx-auto'>
       <PenumbraWaves />