+++ apps/veil/src/pages/tournament/server/previous-epochs.ts
@@ -14,7 +14,6 @@ const DIRECTIONS = ['asc', 'desc'] as const;
 export type PreviousEpochsSortDirection = (typeof DIRECTIONS)[number];
 
 export interface PreviousEpochsRequest {
-  epoch?: number;
   limit: number;
   page: number;
   sortKey: PreviousEpochsSortKey;
@@ -43,7 +42,6 @@ export const getQueryParams = (req: NextRequest): PreviousEpochsRequest => {
 
   const limit = Number(searchParams.get('limit')) || DEFAULT_LIMIT;
   const page = Number(searchParams.get('page')) || 1;
-  const epoch = Number(searchParams.get('epoch')) || undefined;
 
   const sortKeyParam = searchParams.get('sortKey');
   const sortKey =
@@ -58,7 +56,6 @@ export const getQueryParams = (req: NextRequest): PreviousEpochsRequest => {
       : 'desc';
 
   return {
-    epoch,
     limit,
     page,
     sortKey,
@@ -66,12 +63,7 @@ export const getQueryParams = (req: NextRequest): PreviousEpochsRequest => {
   };
 };
 
-const previousEpochsQuery = async ({
-  limit,
-  page,
-  sortDirection,
-  epoch,
-}: PreviousEpochsRequest) => {
+const previousEpochsQuery = async ({ limit, page, sortDirection }: PreviousEpochsRequest) => {
   // 1. Take the 'gauge' table and sort it by 'epoch' and then by 'portion', map asset_id to base64
   const sortedGauge = pindexerDb
     .selectFrom('lqt.gauge as gauge')
@@ -96,7 +88,6 @@ const previousEpochsQuery = async ({
         'gauge',
       ),
     ])
-    .$if(!!epoch, qb => (epoch ? qb.where('epoch', '=', epoch) : qb))
     .orderBy('epoch', sortDirection)
     .groupBy('epoch')
     .execute();
@@ -118,7 +109,6 @@ export async function GET(
   }
 
   const params = getQueryParams(req);
-
   const registryClient = new ChainRegistryClient();
   const [registry, results, total] = await Promise.all([
     registryClient.remote.get(chainId),