+++ apps/veil/src/pages/tournament/ui/lp-rewards.tsx
@@ -1,4 +1,3 @@
-import Link from 'next/link';
 import { useState } from 'react';
 import { observer } from 'mobx-react-lite';
 import { ChevronRight, ExternalLink } from 'lucide-react';
@@ -8,21 +7,132 @@ import { TableCell } from '@penumbra-zone/ui/TableCell';
 import { ValueViewComponent } from '@penumbra-zone/ui/ValueView';
 import { Button } from '@penumbra-zone/ui/Button';
 import { Density } from '@penumbra-zone/ui/Density';
-import { useLpRewards, BASE_LIMIT, BASE_PAGE, Reward } from '../api/use-lp-rewards';
-import { useSortableTableHeaders } from './sortable-table-header';
+import { withdrawPositions } from '@/entities/position/api/withdraw-positions';
+import {
+  useLpRewards,
+  BASE_LIMIT,
+  BASE_PAGE,
+  LpReward,
+} from '@/pages/tournament/api/use-lp-rewards';
+import { useSortableTableHeaders } from '@/pages/tournament/ui/sortable-table-header';
+import { connectionStore } from '@/shared/model/connection';
+import { useRouter } from 'next/navigation';
+import { LpRewardsSortKey } from '../server/lp-rewards';
+import { ValueView, Metadata } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
+import { pnum } from '@penumbra-zone/types/pnum';
+import { useStakingTokenMetadata } from '@/shared/api/registry';
+
+function LoadingRows() {
+  return (
+    <>
+      {new Array(5).map((_, x) => (
+        <div key={x}>
+          <TableCell cell loading>
+            null
+          </TableCell>
+          <TableCell cell loading>
+            null
+          </TableCell>
+          <TableCell cell loading>
+            null
+          </TableCell>
+          <TableCell cell loading>
+            null
+          </TableCell>
+          <TableCell cell loading>
+            null
+          </TableCell>
+        </div>
+      ))}
+    </>
+  );
+}
+
+function LpRewardRow({ lpReward, umMetadata }: { lpReward: LpReward; umMetadata: Metadata }) {
+  const router = useRouter();
+
+  return (
+    <div
+      onClick={() => {
+        router.push(`/inspect/lp/${bech32mPositionId(lpReward.positionId)}`);
+      }}
+      className='grid grid-cols-subgrid col-span-5 hover:bg-action-hoverOverlay transition-colors cursor-pointer'
+    >
+      <TableCell cell>#{lpReward.epoch}</TableCell>
+      <TableCell cell>
+        <div className='max-w-[370px] truncate'>{bech32mPositionId(lpReward.positionId)}</div>
+        <ExternalLink className='size-3 min-w-3 text-neutral-light' />
+      </TableCell>
+      <TableCell cell>
+        <ValueViewComponent
+          valueView={
+            new ValueView({
+              valueView: {
+                case: 'knownAssetId',
+                value: {
+                  amount: pnum(lpReward.rewards).toAmount(),
+                  metadata: umMetadata,
+                },
+              },
+            })
+          }
+          priority='tertiary'
+        />
+      </TableCell>
+      <TableCell cell>
+        {(lpReward.isWithdrawable || lpReward.isWithdrawn) && (
+          <Density slim>
+            <div>
+              <Button
+                priority='primary'
+                disabled={!lpReward.isWithdrawable}
+                onClick={
+                  lpReward.isWithdrawable
+                    ? e => {
+                        e.stopPropagation();
+                        void withdrawPositions([
+                          { position: lpReward.position, id: lpReward.positionId },
+                        ]);
+                      }
+                    : undefined
+                }
+              >
+                {lpReward.isWithdrawn ? 'Withdrawn' : ''}
+                {lpReward.isWithdrawable ? 'Withdraw' : ''}
+              </Button>
+            </div>
+          </Density>
+        )}
+      </TableCell>
+      <TableCell cell>
+        <Density slim>
+          <Button iconOnly icon={ChevronRight}>
+            Go to position information
+          </Button>
+        </Density>
+      </TableCell>
+    </div>
+  );
+}
 
 export const LpRewards = observer(() => {
+  const { subaccount } = connectionStore;
   const [page, setPage] = useState(BASE_PAGE);
   const [limit, setLimit] = useState(BASE_LIMIT);
-  const { getTableHeader, sortBy } = useSortableTableHeaders<keyof Required<Reward>['sort']>();
+  const { getTableHeader, sortBy } = useSortableTableHeaders<keyof LpReward>('epoch');
+  const { data: umMetadata } = useStakingTokenMetadata();
 
-  const {
-    query: { data, isLoading },
-    total,
-  } = useLpRewards(page, limit, sortBy.key, sortBy.direction);
-
-  const loadingArr = new Array(5).fill({ positionId: {} }) as Reward[];
-  const rewards = data ?? loadingArr;
+  const query = useLpRewards(
+    subaccount,
+    page,
+    limit,
+    sortBy.key as LpRewardsSortKey,
+    sortBy.direction,
+  );
+  const { data: queryData, isPending } = query;
+  const { data, total } = queryData ?? { data: [], total: 0 };
+  const loading = isPending;
+  const rewards = data;
 
   const onLimitChange = (newLimit: number) => {
     setLimit(newLimit);
@@ -35,54 +145,23 @@ export const LpRewards = observer(() => {
         <div className='grid grid-cols-[auto_1fr_1fr_100px_48px]'>
           <div className='grid grid-cols-subgrid col-span-5'>
             {getTableHeader('epoch', 'Epoch')}
-            {getTableHeader('positionId', 'Position ID')}
-            {getTableHeader('reward', 'Reward')}
+            <TableCell heading>Position ID</TableCell>
+            {getTableHeader('rewards', 'Reward')}
             <TableCell heading> </TableCell>
             <TableCell heading> </TableCell>
           </div>
 
-          {rewards.map((reward, index) => (
-            <Link
-              href={`/inspect/lp/${isLoading ? index : bech32mPositionId(reward.positionId)}`}
-              key={isLoading ? index : bech32mPositionId(reward.positionId)}
-              className='grid grid-cols-subgrid col-span-5 hover:bg-action-hoverOverlay transition-colors cursor-pointer'
-            >
-              <TableCell cell loading={isLoading}>
-                Epoch #{reward.epoch}
-              </TableCell>
-              <TableCell cell loading={isLoading}>
-                {!isLoading && (
-                  <>
-                    <div className='max-w-full truncate'>
-                      {bech32mPositionId(reward.positionId)}
-                    </div>
-                    <ExternalLink className='size-3 min-w-3 text-neutral-light' />
-                  </>
-                )}
-              </TableCell>
-              <TableCell cell loading={isLoading}>
-                <ValueViewComponent valueView={reward.reward} priority='tertiary' />
-              </TableCell>
-              <TableCell cell loading={isLoading}>
-                <Density slim>
-                  <Button priority='primary' disabled={reward.isWithdrawn}>
-                    {reward.isWithdrawn ? 'Withdrawn' : 'Withdraw'}
-                  </Button>
-                </Density>
-              </TableCell>
-              <TableCell cell loading={isLoading}>
-                <Density slim>
-                  <Button iconOnly icon={ChevronRight}>
-                    Go to position information page
-                  </Button>
-                </Density>
-              </TableCell>
-            </Link>
-          ))}
+          {loading ? (
+            <LoadingRows />
+          ) : (
+            rewards.map((lpReward, index) => (
+              <LpRewardRow key={index} lpReward={lpReward} umMetadata={umMetadata} />
+            ))
+          )}
         </div>
       </Density>
 
-      {!isLoading && total >= BASE_LIMIT && (
+      {!loading && total >= BASE_LIMIT && (
         <Pagination
           totalItems={total}
           visibleItems={rewards.length}