+++ apps/veil/src/pages/tournament/ui/total-delegator-rewards.tsx
@@ -2,79 +2,178 @@ import { observer } from 'mobx-react-lite';
 import { TableCell } from '@penumbra-zone/ui/TableCell';
 import { Density } from '@penumbra-zone/ui/Density';
 import { connectionStore } from '@/shared/model/connection';
-import { usePersonalRewards } from '../api/use-personal-rewards';
-import { ValueView } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
+import { usePersonalRewards, BASE_LIMIT, BASE_PAGE } from '../api/use-personal-rewards';
+import { DelegatorHistorySortKey, LqtDelegatorHistoryData } from '../server/delegator-history';
 import { ValueViewComponent } from '@penumbra-zone/ui/ValueView';
 import { Button } from '@penumbra-zone/ui/Button';
 import { ChevronRight } from 'lucide-react';
 import { useStakingTokenMetadata } from '@/shared/api/registry';
 import { useCurrentEpoch } from '@/pages/tournament/api/use-current-epoch';
-import { pnum } from '@penumbra-zone/types/pnum';
+import { useTournamentSummary } from '../api/use-tournament-summary';
+import { ReactNode, useState } from 'react';
+import { Pagination } from '@penumbra-zone/ui/Pagination';
+import { useSortableTableHeaders } from './sortable-table-header';
+import { LqtSummary } from '@/shared/database/schema';
+import { useGetMetadata } from '@/shared/api/assets';
+import { toValueView } from '@/shared/utils/value-view';
 
-export const VotingRewards = observer(() => {
-  const { subaccount } = connectionStore;
+interface LayoutProps {
+  getTableHeader: (key: 'epoch' | 'reward', label: string) => ReactNode;
+}
 
-  const { epoch, isLoading: epochLoading } = useCurrentEpoch();
-  // This represents a subsequent invocation of `usePersonalRewards` which will use useQuery's cache.
-  const { data, isLoading } = usePersonalRewards(subaccount, epoch, epochLoading);
-  const { data: stakingToken } = useStakingTokenMetadata();
-
-  // TODO: “Cast Vote” still needs a pindexer query—calculate it by dividing the
-  // user’s voting power (from `delegator_history`) by the epoch’s total voting
-  // power (from `delegator_summary`).
-
+const Layout = observer(({ getTableHeader, children }: React.PropsWithChildren<LayoutProps>) => {
   return (
     <Density compact>
       <div className='grid grid-cols-[auto_1fr_1fr_32px]'>
         <div className='grid grid-cols-subgrid col-span-4'>
-          <TableCell heading>Epoch</TableCell>
+          {getTableHeader('epoch', 'Epoch')}
           <TableCell heading>Casted Vote</TableCell>
-          <TableCell heading>Reward</TableCell>
+          {getTableHeader('reward', 'Reward')}
           <TableCell heading> </TableCell>
         </div>
-
-        {data?.data.length
-          ? data.data.map(({ epoch, reward }, index) => {
-              const rewardView = new ValueView({
-                valueView: {
-                  case: 'knownAssetId',
-                  value: {
-                    amount: pnum(reward).toAmount(),
-                    metadata: stakingToken,
-                  },
-                },
-              });
-
-              return (
-                <div key={index} className='grid grid-cols-subgrid col-span-4'>
-                  <TableCell cell loading={isLoading}>
-                    Epoch #{epoch.toString()}
-                  </TableCell>
-
-                  <TableCell cell loading={isLoading}>
-                    -
-                  </TableCell>
-
-                  <TableCell cell loading={isLoading}>
-                    <ValueViewComponent valueView={rewardView} priority='tertiary' />
-                  </TableCell>
-
-                  <TableCell cell loading={isLoading}>
-                    <Density slim>
-                      <Button iconOnly icon={ChevronRight}>
-                        Go to voting reward information
-                      </Button>
-                    </Density>
-                  </TableCell>
-                </div>
-              );
-            })
-          : !isLoading && (
-              <div className='col-span-4 text-sm text-muted-foreground py-4'>
-                No voting rewards found for this account.
-              </div>
-            )}
+        {children}
       </div>
     </Density>
+  );
+});
+
+interface RewardCellProps {
+  reward: LqtDelegatorHistoryData;
+  summary: LqtSummary;
+}
+
+const RewardCell = observer(({ reward, summary }: RewardCellProps) => {
+  const getMetadata = useGetMetadata();
+  const assetId = reward.asset_id;
+  const amount = reward.reward;
+  const metadata = getMetadata(assetId);
+  const valueView = toValueView(metadata ? { metadata, amount } : { assetId, amount });
+  return (
+    <TableCell cell>
+      <span className='font-mono'>{`${((reward.power / summary.total_voting_power) * 100).toFixed(3)}% for `}</span>
+      <ValueViewComponent showValue={false} valueView={valueView} />
+    </TableCell>
+  );
+});
+
+export const VotingRewards = observer(() => {
+  const [page, setPage] = useState(BASE_PAGE);
+  const [limit, setLimit] = useState(BASE_LIMIT);
+  const { getTableHeader, sortBy } = useSortableTableHeaders<DelegatorHistorySortKey>(
+    'epoch',
+    'desc',
+  );
+
+  const { subaccount } = connectionStore;
+
+  const { epoch, status: epochStatus } = useCurrentEpoch();
+  const {
+    query: { status: rewardsStatus },
+    data: rewardsData,
+    total,
+  } = usePersonalRewards(
+    subaccount,
+    epoch,
+    epochStatus !== 'success',
+    page,
+    limit,
+    sortBy.key,
+    sortBy.direction,
+  );
+
+  const { data: stakingToken } = useStakingTokenMetadata();
+
+  // Extract epochs for summary lookup
+  const epochs = [...rewardsData.keys()];
+
+  const { data: rawSummary } = useTournamentSummary(
+    {
+      epochs: epochs.length > 0 ? epochs : undefined,
+    },
+    epochs.length === 0,
+  );
+  if (epoch === undefined || rewardsStatus !== 'success' || rawSummary === undefined) {
+    return (
+      <Layout getTableHeader={getTableHeader}>
+        {new Array(5).map((_, i) => {
+          return (
+            <div key={`loading-${i}`} className='grid grid-cols-subgrid col-span-4'>
+              <TableCell cell loading={true}>
+                undefined
+              </TableCell>
+              <TableCell cell loading={true}>
+                undefined
+              </TableCell>
+              <TableCell cell loading={true}>
+                undefined
+              </TableCell>
+              <TableCell cell loading={true}>
+                <Density slim>
+                  <Button iconOnly icon={ChevronRight} disabled={true}>
+                    Go to voting reward information
+                  </Button>
+                </Density>
+              </TableCell>
+            </div>
+          );
+        })}
+      </Layout>
+    );
+  }
+  const summary = new Map(rawSummary.map(x => [x.epoch, x]));
+
+  const onLimitChange = (newLimit: number) => {
+    setLimit(newLimit);
+    setPage(BASE_PAGE);
+  };
+
+  return (
+    <>
+      <Layout getTableHeader={getTableHeader}>
+        {Array.from(rewardsData.entries(), ([epoch, reward]) => {
+          const matchingSummary = summary.get(epoch);
+          if (!matchingSummary) {
+            throw new Error(`IMPOSSIBLE: no tournament summary at epoch: ${epoch}`);
+          }
+
+          const rewardView = toValueView({ amount: reward.reward, metadata: stakingToken });
+
+          const rowKey = `epoch-${epoch}`;
+
+          return (
+            <div key={rowKey} className='grid grid-cols-subgrid col-span-4'>
+              <TableCell cell>{`Epoch #${epoch}`}</TableCell>
+
+              <RewardCell reward={reward} summary={matchingSummary} />
+
+              <TableCell cell>
+                <ValueViewComponent valueView={rewardView} priority='tertiary' />
+              </TableCell>
+
+              <TableCell cell>
+                <Density slim>
+                  <Button
+                    iconOnly
+                    icon={ChevronRight}
+                    onClick={() => (window.location.href = `/tournament/${epoch}`)}
+                  >
+                    Go to voting reward information
+                  </Button>
+                </Density>
+              </TableCell>
+            </div>
+          );
+        })}
+      </Layout>
+
+      <Pagination
+        totalItems={total}
+        visibleItems={rewardsData.size}
+        value={page}
+        limit={limit}
+        onChange={setPage}
+        onLimitChange={onLimitChange}
+      />
+    </>
   );
 });