+++ flake.nix
@@ -26,11 +26,13 @@
         commonDevPackages = [
           fd
           file
+          firefox
           jq
           just
-          pnpm_10
           nodejs_22
+          pnpm_10
           postgresql
+          libuuid
           wasm-pack
 
           # for deployment/ci
@@ -43,6 +45,8 @@
           export RUST_SRC_PATH=${pkgs.rustPlatform.rustLibSrc} # Required for rust-analyzer
           export NEXT_TELEMETRY_DISABLED=1
           export TURBO_TELEMETRY_DISABLED=1
+          # Override libpath so that builds succeed
+          export LD_LIBRARY_PATH="${pkgs.lib.makeLibraryPath [pkgs.libuuid]}:$LD_LIBRARY_PATH"
         '';
       in
       {