+++ packages/wasm/crate/src/database/mock.rs
@@ -25,6 +25,7 @@ pub fn get_mock_tables() -> Tables {
         auctions: "auctions".to_string(),
         auction_outstanding_reserves: "auction_outstanding_reserves".to_string(),
         tree_commitments: "tree_commitments".to_string(),
+        tree_hashes: "tree_hashes".to_string(),
         tree_last_position: "tree_last_position".to_string(),
         tree_last_forgotten: "tree_last_forgotten".to_string(),
     }