+++ apps/minifront/CHANGELOG.md
@@ -1,5 +1,31 @@
 # minifront
 
+## 6.34.2
+
+### Patch Changes
+
+- Updated dependencies [dc1eb8b]
+- Updated dependencies [f9cd9dd]
+  - @penumbra-zone/protobuf@10.1.0
+  - @penumbra-zone/types@34.0.0
+  - @penumbra-zone/bech32m@17.0.0
+  - @penumbra-zone/client@28.0.0
+  - @penumbra-zone/getters@27.0.0
+  - @penumbra-zone/perspective@57.0.0
+  - @penumbra-zone/ui-deprecated@20.0.0
+  - @penumbra-zone/crypto-web@44.0.0
+
+## 6.34.1
+
+### Patch Changes
+
+- 270def6: feat(minifront): Update v2 header navigation tabs (Portfolio, Shielding), refactor layout to match the minifront v2 layout, and fix active tab highlighting
+- Updated dependencies [085e855]
+  - @penumbra-zone/types@33.1.0
+  - @penumbra-zone/crypto-web@43.0.0
+  - @penumbra-zone/ui-deprecated@19.0.1
+  - @penumbra-zone/perspective@56.0.0
+
 ## 6.34.0
 
 ### Minor Changes