+++ packages/perspective/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @penumbra-zone/perspective
 
+## 57.0.0
+
+### Patch Changes
+
+- Updated dependencies [dc1eb8b]
+- Updated dependencies [f9cd9dd]
+  - @penumbra-zone/protobuf@10.1.0
+  - @penumbra-zone/wasm@48.0.0
+  - @penumbra-zone/bech32m@17.0.0
+  - @penumbra-zone/getters@27.0.0
+
+## 56.0.0
+
+### Patch Changes
+
+- @penumbra-zone/wasm@47.0.0
+
 ## 55.0.0
 
 ### Patch Changes