+++ packages/types/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @penumbra-zone/types
 
+## 34.0.0
+
+### Minor Changes
+
+- dc1eb8b: tct frontier support for freshly generated wallets
+- f9cd9dd: chunk genesis syncing
+
+### Patch Changes
+
+- Updated dependencies [dc1eb8b]
+  - @penumbra-zone/protobuf@10.1.0
+  - @penumbra-zone/bech32m@17.0.0
+  - @penumbra-zone/getters@27.0.0
+
+## 33.1.0
+
+### Minor Changes
+
+- 085e855: use asset id instead of metadata in liquidity tournament idb table
+
 ## 33.0.0
 
 ### Minor Changes